requirements.txt
# Django LIMS Requirements
# Install with: pip install -r requirements.txt

# Core Framework
Django>=5.1

# Image Processing (for barcodes)
Pillow>=10.0

# Barcode Generation
python-barcode>=0.15.1
qrcode>=7.4

# PDF Report Generation
weasyprint>=60.0

# REST API Framework (Phase 1)
djangorestframework>=3.14.0
django-filter>=23.5
drf-spectacular>=0.27.0
django-cors-headers>=4.3.0

# Task Queue & Caching (Phase 1)
celery>=5.3.0
redis>=5.0.0

# Data Import/Export (Phase 1)
openpyxl>=3.1.0

# Instrument Integration (Phase 3)
hl7apy>=1.3.4

# Real-time Communication (Phase 3)
channels>=4.0.0

# Field-level Encryption (Phase 2)
cryptography>=41.0.0

# Optional: For production deployment
# gunicorn>=21.0
# psycopg2-binary>=2.9  # For PostgreSQL
# mysqlclient>=2.2  # For MySQL
